avoid 5-sec dismissal with window.Notification API
After using GH-7 for a week now, I observe that the chrome.notifications
API automatically hides the timer notice after 5 seconds whereas the old
webkit.notifications and the new window.Notification APIs do not.

I prefer the latter behavior because I want the timer notice to be visible
until _I manually_ dismiss it.  For example, I might look away from my
computer for a moment (and thereby miss the timer notice) and return to
using my computer while thinking that the timer is still in progress.

So this commit updates this extension to use the window.Notification API.