Add mute option (#18)
While listening to music the voice or the ringer interrupt sound can really be loud it surprises and flips me down the chair.

Usage: simply go to the control panel (click on the icon), then you will see TTS, ring and a THIRD option that is no sound. Select that last option and the next timer will not have any sound.